+++ tests/aggregated_stream/test_aggregated_stream.py
@@ -1,3 +1,5 @@
+import datetime as dt
+
 from depeche_db import StreamPartitionStatistic
 
 
@@ -20,4 +22,24 @@ def test_stream_statisitics(db_engine, store_with_events, stream_factory):
         next_message_id=account2.events[0].event_id,
         next_message_position=0,
         next_message_occurred_at=account2.events[0].occurred_at,
+        max_position=1,
     )
+
+
+def test_time_to_positions(db_engine, store_with_events, stream_factory):
+    event_store, account, account2 = store_with_events
+    times = [
+        e.get_message_time().replace(tzinfo=dt.timezone.utc)
+        for e in sorted(
+            account.events + account2.events, key=lambda e: e.get_message_time()
+        )
+    ]
+    subject = stream_factory(event_store)
+    subject.projector.update_full()
+
+    for idx, time in enumerate(times):
+        print(time, subject.time_to_positions(time))
+        assert (
+            sum(position for _, position in subject.time_to_positions(time).items())
+            == idx
+        )